Cleaned up tests and logs * broken out matchMedia into its own utility * removed logs * small code cleanup
