+++ polaris-react/src/components/IndexTable/components/Row/tests/Row.test.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import type {ReactElement} from 'react';
 import {mountWithApp} from 'tests/utilities';
 import type {DeepPartial} from '@shopify/useful-types';
+import {setMatchMedia} from 'tests/setup/tests';
 
 import {IndexTable} from '../../../IndexTable';
 import type {IndexTableProps} from '../../../IndexTable';
@@ -15,15 +16,7 @@ import {Checkbox as PolarisCheckbox} from '../../../../Checkbox';
 import styles from '../../../IndexTable.module.scss';
 import type {Range} from '../../../../../utilities/index-provider';
 
-window.matchMedia =
-  window.matchMedia ||
-  function () {
-    return {
-      matches: false,
-      addListener() {},
-      removeListener() {},
-    };
-  };
+setMatchMedia();
 
 const defaultEvent = {
   preventDefault: noop,
